refactor: split big modules + add 500 handler
