+++ src/server/controller.ts
@@ -2,13 +2,13 @@ import type { Db } from "../db/db";
 import type * as T from "../lib/task";
 import type * as E from "../lib/either";
 import type { DbError, ValidationError } from "../errors";
-import type { User } from "../user";
 import { pipe } from "../lib/functions";
 import type * as v from "../lib/validator";
 import * as RA from "../lib/readonlyArray";
 import * as O from "../lib/option";
-import type { EndpointMethod } from "../endpoint";
 import type http from "node:http";
+
+export type EndpointMethod = "GET" | "POST" | "PUT" | "DELETE";
 
 export type ReqValidator<
   M extends EndpointMethod = EndpointMethod,
@@ -24,16 +24,14 @@ export type ReqValidator<
   }
 >;
 
-export function createRouter(
+export function createController<T>(
   routerScheme: ReadonlyArray<
     readonly [
       ReqValidator,
       (args: {
         path: ReadonlyArray<string>;
         request: http.IncomingMessage;
-      }) => (
-        db: Db
-      ) => T.Task<E.Either<ValidationError | DbError, User | User[]>>
+      }) => (db: Db) => T.Task<E.Either<ValidationError | DbError, T | T[]>>
     ]
   >
 ) {
@@ -51,15 +49,14 @@ export function createRouter(
   };
 }
 
-export function route<
-  P extends ReadonlyArray<string>,
-  M extends EndpointMethod
->(
-  validator: ReqValidator<M, P>,
-  handler: (args: {
-    request: http.IncomingMessage & { method: M };
-    path: P;
-  }) => (db: Db) => T.Task<E.Either<ValidationError | DbError, User | User[]>>
-) {
-  return [validator, handler] as const;
+export function routeHandler<T>() {
+  return function <P extends ReadonlyArray<string>, M extends EndpointMethod>(
+    validator: ReqValidator<M, P>,
+    handler: (args: {
+      request: http.IncomingMessage & { method: M };
+      path: P;
+    }) => (db: Db) => T.Task<E.Either<ValidationError | DbError, T | T[]>>
+  ) {
+    return [validator, handler] as const;
+  };
 }
